src/app/samples/sample-playground/sample-playground.spec.ts
import { TestBed, async } from '@angular/core/testing';
import { SamplePlayground } from './sample-playground';
describe('SamplePlayground', () => {
  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [
        SamplePlayground
      ],
    }).compileComponents();
  }));
  it('should create the app', async(() => {
    const fixture = TestBed.createComponent(SamplePlayground);
    const app = fixture.debugElement.componentInstance;
    expect(app).toBeTruthy();
  }));
  it(`should have as title 'app'`, async(() => {
    const fixture = TestBed.createComponent(SamplePlayground);
    const app = fixture.debugElement.componentInstance;
    expect(app.title).toEqual('app');
  }));
  it('should render title in a h1 tag', async(() => {
    const fixture = TestBed.createComponent(SamplePlayground);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('h1').textContent).toContain('Welcome to app!');
  }));
});
